### 2021-04-21 #3195 @arfon
:wave: @ConorMacBride - thanks for your submission to JOSS. can you tell me if this paper is associated with one in a AAS journal?
