### 2021-04-21 #3195 @ConorMacBride
> :wave: @ConorMacBride - thanks for your submission to JOSS. can you tell me if this paper is associated with one in a AAS journal?

It is not.
