### 2021-04-23 #3195 @ConorMacBride
Hi @mbobra, thank you for agreeing to be editor. From the JOSS spreadsheet, the following people (in the order they are listed) have Python as a preferred language and are in the solar physics area: mgalloy, Cadair, hayesla, aasensio, samaloney, dstansby. I have no particular preference. You may know others who would be most suited and available. Thanks!
