### 2021-05-07 #3265 @whedon
[REVIEW]: MCALF: Multi-Component Atmospheric Line Fitting
**Submitting author:** @ConorMacBride (<a href="http://orcid.org/0000-0002-9901-8723">Conor MacBride</a>)
**Repository:** <a href="https://github.com/ConorMacBride/mcalf" target ="_blank">https://github.com/ConorMacBride/mcalf</a>
**Version:** v0.2.1
