### 2021-05-10 ConorMacBride/mcalf#29 @Goobley
The FittingIBIS and LabellingTutorial notebooks expect specific inputs that are not provided
Both of these example notebooks expect multiple specific files that are not provided, and many users will not have anything to hand that can be simply slotted in to replace these. It would be good to have some real data (if only a few pixels) that can work here, simply to allow for ease of testing with real data, and the troubles that typically come with (noise/non-Voigt profiles etc). Since this sort of data is not a great fit for git, it could always be added as a binary file with the release, so as to be downloaded separately if desired.
