### 2021-05-10 #3265 @Goobley
I am quite happy with this software, the level of documentation is excellent (and automated tests are included); the code appears to work well and do "what it says on the tin".

My primary comment has been opened as an issue on the repository (ConorMacBride/mcalf#29):
 - It isn't possible to simply run the examples expecting real data, providing a small amount of this, so that the software can easily be tested on real data (with noise etc), would be very welcomed. Currently I can play with the software using generated profiles, but I can't see if there are difficulties with real data. This would also easily allow testing the parallelisation.

Other points:
- "State of the field", I am not aware of any other packages that propose a toolkit/framework like approach to this problem, or even of automated classification _prior_ to fitting. If this is indeed novel, I would like to see a passing reference to this, explaining why no other software is mentioned with respect to these tasks.
In terms of classification of spectra, Panos et al (doi: 10.3847/1538-4357/aac779) were (to my knowledge) the first to perform autmated spectral classification using machine learning, and whilst k-means was used instead of an MLP, this seems worth mentioning as clustering onto labelled representative profiles could also be used for the classification step.

Overall, I am happy with this work, and would just like to see these minor points addressed.
