### 2021-05-10 #3265 @mgalloy
I would also like to see some data available so that the example notebooks can be executed as is, i.e., the notebook could download the example data from a website if needed. This would make the example notebooks much more useful for new users and demonstrate their capabilities without having to find appropriate data.
