### 2021-05-10 #3265 @ConorMacBride
Thank you @mgalloy and @Goobley for reviewing, and for doing so quickly. Much appreciated.

@Goobley, I've replied in your issue about including example data — I'll check the licence compatibility and include (or link to) a file with some real data. Since the v0.2.0 release, I've added one more example which is in the "latest" documentation (https://mcalf.macbride.me/en/latest/gallery/models/plot_ibis8542model.html), although it is running with randomly generated data.

Regarding "state of field", we are not aware of other similar packages either. I'll update the paper to highlight that it is a novel method. I'll also have a look at Panos et al. (and similar) and include this. I'll provide an update here when I do that.
