### 2021-05-12 ddda5b4
Apply fix in new nbsphinx version
Jinja2 upgrade broke nbsphinx

Signed-off-by: Conor MacBride <conor@macbride.me>
