+++ setup.cfg
@@ -49,7 +49,7 @@ docs =
   sphinx==3.1.2
   sphinx-automodapi==0.12
   ipykernel
-  nbsphinx==0.7.1
+  nbsphinx==0.8.5
   sphinx-rtd-theme==0.5.0
   sphinx-gallery
   pytest
