### 2021-05-13 ConorMacBride/mcalf#29 @ConorMacBride
Hi @Goobley, I've now merged pull requests (#31, #32) where I have added some real IBIS data. I have also updated `LabellingTutorial.ipynb` so it runs as-is, loading the IBIS spectra and selecting a random sample to label using the script. Instead of updating `FittingIBIS.ipynb` to use this data, I created a new gallery example (https://mcalf.macbride.me/en/latest/gallery/models/plot_ibis8542data.html), as the gallery allows for the output of the script to be displayed nicely. The content of `FittingIBIS.ipynb` has mostly been replaced by gallery examples, however, I'd rather not remove it just yet.

As well as the spectral data, I have also added labelled data for training a neural network with, which is also used in the new example. Also included is a file of pre-calculated results from fitting the array of spectra. This lets users quickly see sample results without having to do the computationally intensive task of fitting. I also include the code for doing the fitting. (Pre-calculated results are also *necessary* as the documentation hosting provider would not have the resources to do this much fitting when building the documentation.)

Let me know if you'd like anything more changed here, and feel free to close this issue if you are happy with all of this. Thanks!
