### 2021-05-13 #3265 @ConorMacBride
Hi @mgalloy and @Goobley, I've now added real IBIS data to the repository and updated `LabellingTutorial.ipynb` so it runs as-is with this data. Instead of updating the now mostly redundant `FittingIBIS.ipynb`, I've created a new (more user friendly) gallery example (https://mcalf.macbride.me/en/latest/gallery/models/plot_ibis8542data.html) which covers most of what `FittingIBIS.ipynb` did. This gallery example can also be downloaded and run as-is. I've left a comment in issue https://github.com/ConorMacBride/mcalf/issues/29 with more details. Let me know if you're happy with how I implemented your requests, or if you'd like anything changed. Thanks!
