### 2021-05-13 ConorMacBride/mcalf#29 @Goobley
Hi,
The updated examples look good, thanks for this. However, this comment drew my attention to something I hadn't noticed before:
> Results may differ as there is a random factor when training the neural network.

This should only arise from not being able to fix the initialisation of the MLP in the code (the MLP takes a `random_state` kwarg), and the combination of this, and not having a standardised method for saving/loading the models concerns me. Machine learning _has_ to be reproducible, which I can't see that your classifier model currently is. Retraining something very small like this MLP is fine if the RandomState is fixed so as to make it reproducible.
If I'm missing something here please let me know.
