### 2021-05-13 ConorMacBride/mcalf#29 @ConorMacBride
Hi @Goobley, this is a very good point. The standard way that I (and I think others who use this classifier class) save and reload a trained model is to save it to a python pickle file. For example, to save a trained neural network,

```
import pickle
pkl = open('trainedneuralnetwork.pkl', 'wb')
pickle.dump(model.neural_network, pkl)
pkl.close()
```

And then to load it again later,

```
import pickle
pkl = open('trainedneuralnetwork.pkl', 'rb')
model.neural_network = pickle.load(pkl)  # Overwrite the default untrained model
```

This is covered in `FittingIBIS.ipynb`. I can add this code into the new example, or would you prefer if I allow the `random_state` kwarg to be used with the `IBIS8542Model.train()` method? (I should probably do both.)
