### 2021-05-13 ConorMacBride/mcalf#29 @Goobley
Both of those would be ideal, indeed.

As stated [here](https://scikit-learn.org/stable/modules/model_persistence.html#security-maintainability-limitations), pickle isn't a good long term storage model for networks. I accept that the training data will persist, so allowing the MLP to be seeded correctly to reproduce the output is an important component too.

> In order to rebuild a similar model with future versions of scikit-learn, additional metadata should be saved along the pickled model:
>
>    - The training data, e.g. a reference to an immutable snapshot
>
>    - The python source code used to generate the model
>
>    - The versions of scikit-learn and its dependencies
>
>    - The cross validation score obtained on the training data
