### 2021-05-13 #3265 @mgalloy
I really like the new gallery example at https://mcalf.macbride.me/en/latest/gallery/models/plot_ibis8542data.html.
