### 2021-05-13 34d523c
Test random_state works
Signed-off-by: Conor MacBride <conor@macbride.me>
