+++ setup.cfg
@@ -44,6 +44,7 @@ tests =
   pytest-cov
   pytest-mpl
   tox
+  requests
 docs =
   # Keep these in sync with docs/requirements.txt
   sphinx==3.1.2
