### 2021-05-13 a19af80
Make neural network reproducible, and explain
Signed-off-by: Conor MacBride <conor@macbride.me>
