### 2021-05-13 #3265 @mgalloy
One last thing for me: you mention in the paper the ability to parallelize the code. I don't see anything built into you code to do this (or am I missing it?).
