### 2021-05-13 #3265 @ConorMacBride
Hi @mgalloy, the `IBIS8542Model.fit` method (https://mcalf.macbride.me/en/latest/api/mcalf.models.IBIS8542Model.html#mcalf.models.IBIS8542Model.fit) has an `n_pools` parameter that distributes the list of spectra to fit between a number of parallel processing pools. By default, the spectra are fitted on a single process.
