### 2021-05-13 d8efe7c
Update random_state documentation
Signed-off-by: Conor MacBride <conor@macbride.me>
