### 2021-05-14 b1fa5b9
Use sklearn sample
Signed-off-by: Conor MacBride <conor@macbride.me>
