### 2021-05-14 6d9fde0
Tidy requests dependency
Not needed for testing, but only needed to run this specific example so should not be added as an install dependency of mcalf.

Signed-off-by: Conor MacBride <conor@macbride.me>
