### 2021-05-14 ConorMacBride/mcalf#29 @ConorMacBride
Hi @Goobley, I've now merged a pull request (#33) where I have allowed a `random_state` parameter to be passed through to the scikit-learn MLPClassifier object. I've also updated the example (https://mcalf.macbride.me/en/latest/gallery/models/plot_ibis8542data.html) with details on `random_state` and saving to pickle files. The `random_state` and pickel options are not *perfect* for reproducibility, but until a more robust solution is provided by scikit-learn (i.e. not as dependent on system configurations such as package dependency versions) I guess they will have to do.
