### 2021-05-17 ConorMacBride/mcalf#29 @Goobley
Hi, I'm happy with this now! Just so you know, there's a couple of typos to 'pickel' in the docs on that example. Feel free to close this issue once you've fixed that!
