### 2021-05-17 ConorMacBride/mcalf#29 @ConorMacBride
Excellent! Thanks for spotting the typo. Grepping through the project I could only find one `pickel` so I think I got them all now. Thanks!
