### 2021-05-17 #3265 @ConorMacBride
Thanks @mgalloy and @Goobley for your valuable feedback on this software, helping to improve its quality. The issue (https://github.com/ConorMacBride/mcalf/issues/29) has now been resolved and closed, and it looks like both checklists are now complete @mbobra.
