### 2021-05-17 #3265 @mbobra
@ConorMacBride Thanks! Could you please cut a new release of MCALF and [deposit it in Zenodo](https://guides.github.com/activities/citable-code/)? Please make sure that the author list and title on the Zenodo deposit matches the author list and title on the JOSS paper. Then I can move forward with accepting the paper.
