### 2021-05-17 #3265 @ConorMacBride
Hi @mbobra, excellent! I've now created a new release ([MCALF v0.2.1](https://github.com/ConorMacBride/mcalf/releases/tag/v0.2.1)) and a new version in Zenodo ([10.5281/zenodo.4767888](http://doi.org/10.5281/zenodo.4767888)). Thanks!
