### 2021-05-17 #3265 @kyleniemeyer
@ConorMacBride Hi there, I'm the AEIC on duty this week doing some final checks before publishing.

I only noticed a very minor issue: in the Panos et al. reference, it looks like some of the words in the title need to be wrapped with `{}` to preserve capitalization. Could you fix that and let me know?
