### 2021-05-17 #3265 @ConorMacBride
Hi @kyleniemeyer, I'll fix that now. Should I also preserve the all-caps title in Felipe et al. (https://doi.org/10.1088/0004-637x/795/1/9)?
