### 2021-05-17 #3265 @kyleniemeyer
@ConorMacBride no, only words that should actually be capitalized in a regular sentence, I think.
