### 2021-05-17 #3265 @ConorMacBride
Thanks @kyleniemeyer. I've now merged a PR (https://github.com/ConorMacBride/mcalf/pull/35) which keeps the `Mg` capitalised and sets the `ii` to small caps.
